final mikey pull request
I cleaned up unused code.
I got rid of the overlay tab.
I got the side by side to have the RSA feed functional.
