+++ main.py
@@ -4,18 +4,12 @@ __credits__ = ["Addison Raak", "Michael Antkiewicz", "Ka'ulu Ng", "Nicholas Bald
 
 
 import matplotlib
-import serial
+import sys
 matplotlib.use('Qt5Agg')
 
 from PyQt5 import QtWidgets
-from data_stream import *
-from graph_widget import *
 from Side_By_Side import Side_By_Side_Tab
 from Overlay import Overlay_Tab
-
-# change this port number based on the machine you are using
-# will need to get changed if the USB port changes
-#arduinoData = serial.Serial("/dev/cu.usbmodem145401", 9600)
 
 
 # Class for the main widget of the program, which will have the tab manager
@@ -31,10 +25,12 @@ class TabManager(QtWidgets.QWidget):
         tab_controller = QtWidgets.QTabWidget()
 
         self.side_tab = Side_By_Side_Tab()
+
+        # overlay was not finished due to threading
         self.overlay_tab = Overlay_Tab()
 
         # make the two tabs needed (overlay and sidebyside)
-        tab_controller.addTab(self.overlay_tab, "Overlay")
+        #tab_controller.addTab(self.overlay_tab, "Overlay")
         tab_controller.addTab(self.side_tab, "Side by Side")
 
         tab_container.addWidget(tab_controller)
